requirements.txt
aiohttp==3.6.2
argon2-cffi==20.1.0
args==0.1.0
asgiref==3.2.7
astroid @ file:///D:/bld/astroid_1588698392162/work
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
brotlipy==0.7.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
clint==0.5.1
colorama==0.4.3
conda==4.8.3
conda-package-handling==1.6.0
coverage==5.1
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
discord.py==1.3.3
Django==3.0.8
django-admin-bootstrapped==2.5.7
django-admin-interface==0.12.2
django-admin-rangefilter==0.6.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-braces==1.14.0
django-colorfield==0.3.2
django-cors-middleware==1.5.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-filters==0.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.2.0
django-jet==1.0.8
django-pandas==0.6.2
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django.js==0.8.1
emoji==0.5.4
et-xmlfile==1.0.1
Faker==4.0.3
get==2019.4.13
idna==2.9
install==1.3.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lml==0.0.9
mamba==0.11.0
MarkupPy==1.14
mccabe==0.6.1
menuinst==1.4.16
misaka==2.1.1
multidict==4.7.6
numpy==1.19.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.5
phonenumbers==8.12.5
Pillow==7.1.2
post==2019.4.13
psycopg2-binary==2.8.5
public==2019.4.13
pycosat==0.6.3
pycparser==2.20
pycryptodome==3.9.7
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
pylint==2.5.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
ruamel-yaml==0.15.80
sage==0.0.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
to==0.3
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1589469402899/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1591181521996/work
urllib3==1.25.9
websockets==8.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
